refactored main.cpp and moved player, board etc. into seperate files
