+++ Snake/input.h
@@ -1,0 +1,18 @@
+#pragma once
+
+namespace Input {
+    enum Action {
+        Up,
+        Left,
+        Down,
+        Right,
+        Quit,
+    };
+
+    void setNonBlockingInput(bool enable);
+
+    void inputPooling(Action& input);
+}
+
+
+
