+++ Snake/types.h
@@ -1,0 +1,22 @@
+#pragma once
+#include <string>
+
+
+namespace Types {
+    struct Position {
+        int x {};
+        int y {};
+
+        Position& operator+=(const Position& other) {
+            x += other.x;
+            y += other.y;
+            return *this;
+        }
+    };
+
+
+    struct Tile {
+        std::string icon {" "};
+        Position pos {};
+    };
+}
